Add a variable scope for the tensor Hx_plain_gradients
The tensor Hx_plain_gradients is defined in PerlmutterHvp but is not
created there. It's called in another file using a lamda function, so
the enclosing_scope function won't fully cover that scenario.
That's why the complete scope is passed as a parameter in the function
where Hx_plain_gradients is defined, so the scope can be reestablished
once the gradients are created.